Create robot
Write the function robot that moves a robot from one location to another. The function robot consumes four values: the initial robot position as a Posn, the initial direction the robot is facing (one of the four symbols ’north, ’east, ’south, or ’west), a turning action for the robot (one of the three symbols ’left, ’right, ’noturn) and a non-negative distance for the robot to move. On a left turn, the robot turns counterclockwise 90 degrees. On a right turn the robot turns clockwise 90 degrees. On no turn, the robot does not turn. After performing the turning action, the robot then moves forward the given distance to its new position. The function should produce a new Posn that records the robots new position after turning and moving the given from its initial position and orientation (you should assume that ’east is the positive x-direction and ’north is the positive y-direction, with ’west and ’south being the corresponding negative directions).