Create student
Using pure structural recursion: Write the function tallest which consumes a list of symbols (representing first names of students) and a list of positive numbers (representing the corresponding height of the students). The function should produce the name of the student who is the tallest.

Using accumulative recursion: Write the function tallest which consumes a list of symbols (representing first names of students) and a list of positive numbers (representing the corresponding height of the students).he function should produce the name of the student who is the shortest.

Write the function student-al which consumes a list of symbols (representing ﬁrst names of students) and a list of positive numbers (representing the corresponding height of the students). The function student-al should produce an association list (i.e., a list of lists) with the keys being the student names (symbols) and the values being the height of the corresponding student (positive numbers).

Write the function basketball which consumes an association list as created in the previous part (i.e., a list of lists, with the keys being the names as symbols, and the values being the numerical height of the students) and a positive height and produces the list of names (i.e., symbols) for those students that are at least as tall as the given height, in the same order that the names appear in the consumed list.