Create plants
Define a structure PlantInfo. The first field is called name, and it gives the name of the plant. The second field is called zone, and it ranges from 0 to 9 inclusive. The third field is called subzone, and is one of ’a or ’b.

Write a function, find-hardy-plants, that consumes a CityInfo value and a PlantInfoList value, in that order. The function produces a PlantInfoList value containing only PlantInfos for plants that will grow in the given zone. For example, based on the sample list above, (find-hardy-plants (make-cityinfo "Edmonton" 3 ’a) sample-plant-data) produces a one-element list while (find-hardy-plants (make-cityinfo "Quebec" 4 ’b) sample-plant-data) produces a three-element list. 

Write a function, find-growing-cities, that consumes a PlantInfo and a CityInfoList in that order, and produces a (listof Str) containing the names of all of the cities in which the plant will grow. The order of the names in the input list must be maintained in the output list.
 
Suppose an online plant-selling business wants to ensure that its catalogue has plants suitable for all the cities in Canada. Write a function, find-plantless-cities, that consumes a PlantInfoList and a CityInfoList in that order, and produces a CityInfoList containing all of the CityInfos in the given CityInfoList for which no plants in the PlantInfoList would survive the winter.