Create int-lst
Write a function elements-more-than that consumes a (listof Int) and an Int in that order. The function will produce a list of integers from the original list that are strictly greater than the integer given. The order of the integers in the produced list should match the order of the integers in the list consumed.


Write a function arithmetic-sequence? that consumes a (listof Int) and produces true if the list is an arithmetic sequence and false otherwise. If the list consumed is empty or has only one or two elements, the function should produce true.

Write a function digits->integer that consumes a (listof Int) and produces the integer represented by those digits, where the least significant digit is at the beginning of the list and the most significant digit is at the end of the list. If any of the values in the list are not single digit, non-negative integers, then the function should produce ’error. If the list consumed is empty, then the function should produce 0.