Create map-fn
Write a function mapfn which consumes a list of functions (each of which takes two numbers as arguments) and a list of two numbers. It should produce the list of the results of applying each function in turn to the given two numbers.

Write a predicate function is-in-order? that consumes a list of (predicate function, binary relational operator) pairs and a list of operands. A pair in the first list consists of a predicate to determine the data type of the list of operands and a binary relational operator that consumes the same data type. The first list is a non-empty list and can contain at most one such pair from any data type. Each element of the second list is of the same data type. The second list may be empty.