Create cards
Write a function better-card which consumes two cards, and produces the Card which is the better of the two. The better card is either the card with the better suit (the suits are increasing from ’clubs (worst), ’diamonds (second-worst), ’hearts (second-best), or ’spades (best)) or, if the cards have the same suit, then the best card is the card with the largest/highest rank. If the two consumed Cards are the same (i.e., same suit and rank), return either one.

Write the function hand-value which consumes a Hand and produces a symbol indicating the best hand-value of the given Hand (where the arrangement/order of the three cards in the Hand can be altered).