Create alfs
Construct the following functions using abstract list functions:

The function x-coords-of-posns, which consumes a list of anything at all, and produces the list of all of the x-coordinates of the Posns in the list, in the same order which they occur in the consumed list.

The function alternating-sum, which consumes a list of numbers (list a1 a2 ... an) and produces the alternating sum a1 −a2 +a3 −a4 +···(−1)n−1an; if the list is empty, produce 0.

The function remove-duplicates, which consumes a list of numbers, and produces the list with all but the first occurrence of every number removed. For example, (remove-duplicates ’(1 4 2 1 5 4)) => ’(1 4 2 5)

The function first-col which consumes a list of non-empty list of numbers, denoting a rectangular matrix of numbers, and produces the first column of the matrix, as a list. For example, (first-col ’((1 2 3 4) (5 6 7 8) (9 10 11 12))) => (list 1 5 9)

The function add1-mat which consumes a (listof (listof Num)), denoting a rectangular matrix of numbers, and produces the matrix resulting from adding 1 to every entry of the matrix.
For example, (add1-mat ’((1 2 3 4) (5 6 7 8) (9 10 11 12))) =>(list (list 2 3 4 5) (list 6 7 8 9) (list 10 11 12 13))

The function sum-at-zero, which consumes a list of functions (f1,..., fn) (where each consumes and produces a number), and produces the value f1(0)+···+ fn(0).