Create bst
Submit the function root-at-smallest. Although it is tricky to think about, the function itself is quite simple. (The sample solution is less than 10 lines long.) Hint: use local definitions to identify in your code the significant parts of Figure 1; i.e. x, y, A, and B.

The reason the above root-at-smallest function is useful is that it helps us implement bst-remove. bst-remove consumes a key (which is a Num) and a BST, and produces the BST with the node containing the given key removed. (Produce the original tree unchanged if the key is not present.)